third_party/blink/web_tests/platform/generic/http/tests/loading/nested_bad_objects-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
 PASS - nested image objects with bad mimetype do not cause a crash.
